NGEN-9 Refactor regular utilities place
