+++ src/app/core/models/regular-string.ts
@@ -1,74 +1,7 @@
-import { replaceLetter } from "@ngen-core/functions";
-import { LetterUtils, RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
+import { RegularUtils } from "@ngen-generation/generator-algorithms/letter-finalization/utils";
 import { RegularCharacter } from "./regular-character";
 
 export class RegularString {
-   private static deleteNonRegulars(input: string): string {
-      let output = "";
-      for (let i = 0; i < input.length; i++) {
-         if (LetterUtils.is("letter", input[i]) || RegularUtils.isRegular(input[i])) {
-            output += input[i];
-         } else {
-            output += RegularUtils.symbols.wildcard;
-         }
-      }
-      return output;
-   }
-
-   private static fixReferences(input: string): string {
-      let referenceRegular = "";
-      const referencingSets: Set<number>[] = [];
-
-      for (let i = 0; i < input.length; i++) {
-         referencingSets[i] = new Set();
-      }
-      for (let i = 0; i < input.length; i++) {
-         if (RegularUtils.isReference(input[i])) {
-            const referenceTo = Number(input[i]);
-            if (referenceTo === i || referenceTo >= input.length) {
-               referenceRegular += RegularUtils.symbols.wildcard;
-               continue;
-            } else {
-               if (referenceTo > i) {
-                  referencingSets[referenceTo].add(i);
-               } else {
-                  referencingSets[i].add(referenceTo);
-               }
-            }
-         }
-         referenceRegular += input[i];
-      }
-
-      for (let i = referencingSets.length - 1; i >= 0; i--) {
-         if (referencingSets[i].size <= 1) {
-            continue;
-         }
-         const references = referencingSets[i];
-         const nextReferenceInChain = Math.max(...Array.from(references));
-         referencingSets[i] = new Set([nextReferenceInChain]);
-         references.delete(nextReferenceInChain);
-         references.forEach(ref => referencingSets[nextReferenceInChain].add(ref));
-      }
-
-      const referencingNumbers: (number | undefined)[] = referencingSets.map(references => Array.from(references)[0]);
-      let output = "";
-      for (let i = 0; i < referenceRegular.length; i++) {
-         if (RegularUtils.isReference(referenceRegular[i])) {
-            output += referencingNumbers[i] ?? RegularUtils.symbols.wildcard;
-         } else if (referencingNumbers[i] === undefined) {
-            output += referenceRegular[i];
-         } else {
-            const letterToReference = referenceRegular[i];
-            let idx = i;
-            while (referencingNumbers[idx] !== undefined) {
-               idx = referencingNumbers[idx]!;
-            }
-            output = replaceLetter(output, idx, letterToReference) + referencingNumbers[i];
-         }
-      }
-      return output;
-   }
-
    private readonly characters: RegularCharacter[] = [];
 
    constructor(str?: string) {
@@ -76,7 +9,7 @@ export class RegularString {
          return;
       }
 
-      this.characters.push(...this.parseString(str));
+      this.characters.push(...RegularUtils.parseStringToRegularCharacters(str));
    }
 
    public append(str: string | RegularString): void {
@@ -140,11 +73,5 @@ export class RegularString {
 
    public clone(): RegularString {
       return new RegularString(this.toString());
-   }
-
-   private parseString(str: string): RegularCharacter[] {
-      return RegularString.deleteNonRegulars(RegularString.fixReferences(str.toLowerCase()))
-         .split("")
-         .map(char => new RegularCharacter(char));
    }
 }
